Initial subqueries support API

+++ MysqliDb.php
@@ -94,13 +94,19 @@ class MysqliDb
     protected $port;
 
     /**
+     * Is Subquery object
+     *
+     */
+    protected $isSubQuery = false;
+
+    /**
      * @param string $host
      * @param string $username
      * @param string $password
      * @param string $db
      * @param int $port
      */
-    public function __construct($host, $username, $password, $db, $port = NULL)
+    public function __construct($host = NULL, $username = NULL, $password = NULL, $db = NULL, $port = NULL)
     {
         $this->host = $host;
         $this->username = $username;
@@ -110,9 +116,14 @@ class MysqliDb
             $this->port = ini_get ('mysqli.default_port');
         else
             $this->port = $port;
-        
-        $this->connect();
         $this->setPrefix();
+
+        if ($host == null && $username == null && $db == null) {
+            $this->isSubQuery = true;
+            return;
+        }
+        // for subqueries we do not need database connection and redefine root instance
+        $this->connect();
         self::$_instance = $this;
     }
 
@@ -122,6 +133,9 @@ class MysqliDb
      */
     public function connect()
     {
+        if ($this->isSubQuery)
+            return;
+
         $this->_mysqli = new mysqli ($this->host, $this->username, $this->password, $this->db, $this->port)
             or die('There was a problem connecting to the database');
 
@@ -165,6 +179,7 @@ class MysqliDb
     public function setPrefix($prefix = '')
     {
         $this->_prefix = $prefix;
+        return $this;
     }
 
     /**
@@ -233,6 +248,10 @@ class MysqliDb
         $column = is_array($columns) ? implode(', ', $columns) : $columns; 
         $this->_query = "SELECT $column FROM $this->_prefix$tableName";
         $stmt = $this->_buildQuery($numRows);
+
+        if ($this->isSubQuery)
+            return $this;
+
         $stmt->execute();
         $this->_stmtError = $stmt->error;
         $this->reset();
@@ -250,6 +269,10 @@ class MysqliDb
     public function getOne($tableName, $columns = '*') 
     {
         $res = $this->get ($tableName, 1, $columns);
+
+        if (is_object($res))
+            return $res;
+
         if (isset($res[0]))
             return $res[0];
 
@@ -265,6 +288,9 @@ class MysqliDb
      */
     public function insert($tableName, $insertData)
     {
+        if ($this->isSubQuery)
+            return;
+
         $this->_query = "INSERT into $this->_prefix$tableName";
         $stmt = $this->_buildQuery(null, $insertData);
         $stmt->execute();
@@ -284,6 +310,9 @@ class MysqliDb
      */
     public function update($tableName, $tableData)
     {
+        if ($this->isSubQuery)
+            return;
+
         $this->_query = "UPDATE $this->_prefix$tableName SET ";
 
         $stmt = $this->_buildQuery(null, $tableData);
@@ -304,6 +333,9 @@ class MysqliDb
      */
     public function delete($tableName, $numRows = null)
     {
+        if ($this->isSubQuery)
+            return;
+
         $this->_query = "DELETE FROM $this->_prefix$tableName";
 
         $stmt = $this->_buildQuery($numRows);
@@ -488,6 +520,21 @@ class MysqliDb
         array_push ($this->_bindParams, $value);
     }
 
+    protected function _buildPair ($operator, $value) {
+        if (!is_object($value)) {
+            $this->_bindParam ($value);
+            $comparison = ' ' . $operator. ' ? ';
+            return $comparison;
+        }
+
+        $subQuery = $value->getSubQuery();
+        $comparison = " " . $operator . " (" . $subQuery['query'] . ")";
+        foreach ($subQuery['params'] as $v)
+            $this->_bindParam ($v);
+
+        return $comparison;
+    }
+
     /**
      * Abstraction method that will compile the WHERE statement,
      * any passed update data, and the desired rows.
@@ -582,9 +629,13 @@ class MysqliDb
                         case 'not in':
                         case 'in':
                             $comparison = ' ' . $key . ' (';
-                            foreach($val as $v){
-                                $comparison .= ' ?,';
-                                $this->_bindParam ($v);
+                            if (is_object ($val)) {
+                                $comparison .= $this->_buildPair ("", $val);
+                            } else {
+                                foreach ($val as $v) {
+                                    $comparison .= ' ?,';
+                                    $this->_bindParam ($v);
+                                }
                             }
                             $comparison = rtrim($comparison, ',').' ) ';
                             break;
@@ -596,14 +647,12 @@ class MysqliDb
                             break;
                         default:
                             // We are using a comparison operator with only one parameter after it
-                            $comparison = ' '.$key.' ? ';
-                            $this->_bindParam ($val);
+                            $comparison = $this->_buildPair ($key, $val);
                     }
                 } else if ($value[1] === null) {
                     $comparison = '';
                 } else {
-                    $comparison = ' = ? ';
-                    $this->_bindParam ($value[1]);
+                    $comparison = $this->_buildPair ("=", $value[1]);
                 }
                 $this->_query .= $column.$comparison;
             }
@@ -637,6 +686,11 @@ class MysqliDb
                 $this->_query .= ' LIMIT ' . (int)$numRows;
         }
 
+        $this->_lastQuery = $this->replacePlaceHolders($this->_query, $this->_bindParams);
+
+        if ($this->isSubQuery)
+            return;
+
         // Prepare query
         $stmt = $this->_prepareQuery();
 
@@ -644,7 +698,6 @@ class MysqliDb
         if (count ($this->_bindParams) > 1)
             call_user_func_array(array($stmt, 'bind_param'), $this->refValues($this->_bindParams));
 
-        $this->_lastQuery = $this->replacePlaceHolders($this->_query, $this->_bindParams);
         return $stmt;
     }
 
@@ -708,7 +761,8 @@ class MysqliDb
      */
     public function __destruct()
     {
-        $this->_mysqli->close();
+        if (!$this->isSubQuery)
+            $this->_mysqli->close();
     }
 
     /**
@@ -741,7 +795,10 @@ class MysqliDb
         $newStr = "";
 
         while ($pos = strpos ($str, "?")) {
-            $newStr .= substr ($str, 0, $pos) . $vals[$i++];
+            $val = $vals[$i++];
+            if (is_object ($val))
+                $val = '[object]';
+            $newStr .= substr ($str, 0, $pos) . $val;
             $str = substr ($str, $pos + 1);
         }
         return $newStr;
@@ -763,6 +820,18 @@ class MysqliDb
      */
     public function getLastError () {
         return $this->_stmtError . " " . $this->_mysqli->error;
+    }
+
+    public function getSubQuery () {
+        if (!$this->isSubQuery)
+            return null;
+
+        array_shift ($this->_bindParams);
+        $val = Array ('query' => $this->_query,
+                      'params' => $this->_bindParams
+                );
+        $this->reset();
+        return $val;
     }
 
     /* Helper functions */
@@ -839,6 +908,14 @@ class MysqliDb
      */
     public function func ($expr, $bindParams = null) {
         return Array ("[F]" => Array($expr, $bindParams));
+    }
+
+    /**
+     * Method creates new mysqlidb object for a subquery generation
+     */
+    public static function subQuery()
+    {
+        return new MysqliDb();
     }
 
 } // END class
